// goldex-backend/src/wallet/enum/transaction.type.enum.ts
export enum TransactionTypeEnum {
  DEPOSIT = "DEPOSIT",
  WITHDRAWAL = "WITHDRAWAL",
  BUY = "BUY",
  SELL = "SELL",
  ADMIN_ADJUSTMENT = "ADMIN_ADJUSTMENT",
  FEE = "FEE",
  REFERRAL = "REFERRAL",
  ORDER = "ORDER",
  ORDER_CANCEL = "ORDER_CANCEL",
  ORDER_REJECTED = "ORDER_REJECTED",
  MATERIAL_DEPOSIT = "MATERIAL_DEPOSIT",
  MATERIAL_WITHDRAW = "MATERIAL_WITHDRAW",
  CREDIT_DEPOSIT = "CREDIT_DEPOSIT",
  CREDIT_WITHDRAWAL = "CREDIT_WITHDRAWAL",
  CREDIT_LIQUIDATION = "CREDIT_LIQUIDATION",
  CREDIT_SETTLEMENT = "CREDIT_SETTLEMENT",
  MATERIAL_FREEZE = "MATERIAL_FREEZE",
  MATERIAL_UNFREEZE = "MATERIAL_UNFREEZE",
  P2P_WITHDRAW_LOCK = "P2P_WITHDRAW_LOCK",
  P2P_WITHDRAW_SETTLE = "P2P_WITHDRAW_SETTLE",
  P2P_WITHDRAW_RELEASE = "P2P_WITHDRAW_RELEASE",
  P2P_DEPOSIT_SETTLE = "P2P_DEPOSIT_SETTLE",
}